tets

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -8,14 +8,17 @@ from model_state import Base, State
 from sqlalchemy import (create_engine)
 from sqlalchemy.orm import sessionmaker
 if __name__ == "__main__":
-
-    engine = create_engine('mysql+pymysql://{}:{}@localhost:3306/{}'
-                           .format(sys.argv[1], sys.argv[2], sys.argv[3]))
+    user = sys.argv[1]
+    password = sys.argv[2]
+    db_name = sys.argv[3]
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'
+                           .format(user, password, db_name))
     Base.metadata.create_all(engine)
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    for state in session.query(State).filter(State.name.like('%a%')).all():
-        session.delete(state)
-        
+    for state in session.query(State):
+        if 'a' in state.name:
+           session.delete(state)
+
     session.commit()
